+++ test/integration/test_resource_manager_v2.py
@@ -26,17 +26,9 @@ import datetime
 import random
 from ibm_cloud_sdk_core import *
 from ibm_platform_services.resource_manager_v2 import *
-from dotenv import load_dotenv
 
-# Read config file
-configFile = 'resource_manager.env'
-configLoaded = None
-
-if os.path.exists(configFile):
-    load_dotenv(dotenv_path=configFile)
-    configLoaded = True
-else:
-    print('External configuration was not found, skipping tests...')
+# Location of our config file.
+config_file = 'resource_manager.env'
 
 class TestResourceManagerV2(unittest.TestCase):
     """
@@ -45,22 +37,26 @@ class TestResourceManagerV2(unittest.TestCase):
 
     @classmethod
     def setUpClass(cls):
-        if not configLoaded:
-            raise unittest.SkipTest(
-                'External configuration not available, skipping...')
+        if os.path.exists(config_file):
+            os.environ['IBM_CREDENTIALS_FILE'] = config_file
 
-        # Construct the first service instance.
-        cls.service1 = ResourceManagerV2.new_instance(service_name='RMGR1')
-        assert cls.service1 is not None
+            cls.config = read_external_sources(
+                ResourceManagerV2.DEFAULT_SERVICE_NAME
+            )
+            assert cls.config is not None
 
-        # Construct the second service instance.
-        cls.service2 = ResourceManagerV2.new_instance(service_name='RMGR2')
-        assert cls.service2 is not None
+            # Construct the first service instance.
+            cls.service = ResourceManagerV2.new_instance(service_name=ResourceManagerV2.DEFAULT_SERVICE_NAME)
+            assert cls.service is not None
 
-        # setup default values
-        cls.test_quota_id = '7ce89f4a-4381-4600-b814-3cd9a4f4bdf4'
-        cls.test_user_account_id = '60ce10d1d94749bf8dceff12065db1b0'
-        cls.new_resource_group_id = ''
+            # Construct the second service instance.
+            cls.alt_service = ResourceManagerV2.new_instance(service_name='ALT_RESOURCE_MANAGER')
+            assert cls.alt_service is not None
+
+            # setup default values
+            cls.test_quota_id = cls.config['QUOTA_ID']
+            cls.test_user_account_id = cls.config['USER_ACCOUNT_ID']
+            cls.new_resource_group_id = ''
 
         print('\nSetup complete.')
 
@@ -70,10 +66,10 @@ class TestResourceManagerV2(unittest.TestCase):
         print('\nClean up complete.')
 
     def test_00_check_service(self):
-        assert self.service1 is not None
+        assert self.service is not None
 
     def test_01_list_quota_definitions(self):
-        response = self.service1.list_quota_definitions()
+        response = self.service.list_quota_definitions()
         assert response is not None
         assert response.get_status_code() == 200
 
@@ -83,7 +79,7 @@ class TestResourceManagerV2(unittest.TestCase):
         assert resources is not None
 
     def test_02_get_quota_definition(self):
-        response = self.service1.get_quota_definition(id=self.test_quota_id)
+        response = self.service.get_quota_definition(id=self.test_quota_id)
         assert response is not None
         assert response.get_status_code() == 200
 
@@ -91,7 +87,7 @@ class TestResourceManagerV2(unittest.TestCase):
         assert result is not None
 
     def test_03_list_resource_groups_in_an_account(self):
-        response = self.service1.list_resource_groups(
+        response = self.service.list_resource_groups(
             account_id=self.test_user_account_id)
         assert response is not None
         assert response.get_status_code() == 200
@@ -114,7 +110,7 @@ class TestResourceManagerV2(unittest.TestCase):
         assert resources.get('updated_at') is not None
 
     def test_04_create_resource_group_in_an_account(self):
-        response = self.service1.create_resource_group(
+        response = self.service.create_resource_group(
             name='TestGroup', account_id=self.test_user_account_id)
         assert response is not None
         assert response.get_status_code() == 201
@@ -126,7 +122,7 @@ class TestResourceManagerV2(unittest.TestCase):
         self.__class__.new_resource_group_id = result.get('id')
 
     def test_05_get_resource_group_by_id(self):
-        response = self.service1.get_resource_group(
+        response = self.service.get_resource_group(
             id=self.new_resource_group_id)
         assert response is not None
         assert response.get_status_code() == 200
@@ -135,7 +131,7 @@ class TestResourceManagerV2(unittest.TestCase):
         assert result is not None
 
     def test_06_update_resource_group_by_id(self):
-        response = self.service1.update_resource_group(
+        response = self.service.update_resource_group(
             id=self.new_resource_group_id, name='TestGroup2', state='ACTIVE')
         assert response is not None
         assert response.get_status_code() == 200
@@ -144,7 +140,7 @@ class TestResourceManagerV2(unittest.TestCase):
         assert result is not None
 
     def test_07_delete_resource_group_by_id(self):
-        response = self.service2.delete_resource_group(
+        response = self.alt_service.delete_resource_group(
             id=self.new_resource_group_id)
         assert response is not None
         assert response.get_status_code() == 204